Renamed to match other interfaces.
