+++ app/src/main/java/com/daniellegolinsky/funshine/di/ApplicationModule.kt
@@ -10,8 +10,8 @@ import androidx.datastore.preferences.preferencesDataStoreFile
 import com.daniellegolinsky.funshine.api.OpenMeteoWeatherService
 import com.daniellegolinsky.funshine.datastore.IWeatherSettingsDataStore
 import com.daniellegolinsky.funshine.datastore.WeatherSettingsDataStore
+import com.daniellegolinsky.funshine.utilities.IResourceProvider
 import com.daniellegolinsky.funshine.utilities.ResourceProvider
-import com.daniellegolinsky.funshine.utilities.ResourceProviderImpl
 import com.google.gson.GsonBuilder
 import dagger.Module
 import dagger.Provides
@@ -98,7 +98,7 @@ object ApplicationModule {
 
     @Provides
     @Singleton
-    fun providesResourceProvider(@ApplicationContext context: Context): ResourceProvider {
-        return ResourceProviderImpl(context)
+    fun providesResourceProvider(@ApplicationContext context: Context): IResourceProvider {
+        return ResourceProvider(context)
     }
 }
