+++ app/src/main/java/com/daniellegolinsky/funshine/utilities/IResourceProvider.kt
@@ -1,0 +1,7 @@
+package com.daniellegolinsky.funshine.utilities
+
+interface IResourceProvider {
+    fun getString(id: Int): String
+    fun getString(id: Int, formatString: String): String
+    fun getString(id: Int, formatInt: Int): String
+}
