+++ app/src/main/java/com/daniellegolinsky/funshine/utilities/ResourceProvider.kt
@@ -1,7 +1,22 @@
 package com.daniellegolinsky.funshine.utilities
 
-interface ResourceProvider {
-    fun getString(id: Int): String
-    fun getString(id: Int, formatString: String): String
-    fun getString(id: Int, formatInt: Int): String
+import android.content.Context
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+
+class ResourceProvider @Inject constructor(
+    @ApplicationContext private val context: Context
+) : IResourceProvider {
+    override fun getString(id: Int): String {
+        return context.getString(id)
+    }
+
+    override fun getString(id: Int, formatString: String) : String {
+        return context.getString(id, formatString)
+    }
+
+    override fun getString(id: Int, formatInt: Int) : String {
+        return context.getString(id, formatInt)
+    }
 }
+
