--- app/src/main/java/com/daniellegolinsky/funshine/utilities/ResourceProviderImpl.kt
@@ -1,22 +1,0 @@
-package com.daniellegolinsky.funshine.utilities
-
-import android.content.Context
-import dagger.hilt.android.qualifiers.ApplicationContext
-import javax.inject.Inject
-
-class ResourceProviderImpl @Inject constructor(
-    @ApplicationContext private val context: Context
-) : ResourceProvider {
-    override fun getString(id: Int): String {
-        return context.getString(id)
-    }
-
-    override fun getString(id: Int, formatString: String) : String {
-        return context.getString(id, formatString)
-    }
-
-    override fun getString(id: Int, formatInt: Int) : String {
-        return context.getString(id, formatInt)
-    }
-}
-
