+++ app/src/test/java/com/daniellegolinsky/funshine/testImplementations/MockResourceProvider.kt
@@ -1,8 +1,8 @@
 package com.daniellegolinsky.funshine.testImplementations
 
-import com.daniellegolinsky.funshine.utilities.ResourceProvider
+import com.daniellegolinsky.funshine.utilities.IResourceProvider
 
-class MockResourceProvider: ResourceProvider {
+class MockResourceProvider: IResourceProvider {
     override fun getString(id: Int): String {
         return "You requested ID: $id"
     }
